@Patryk Milewski suggestions
Took 11 minutes

+++ ServerHttp/src/main/java/io/raspberrywallet/contract/Manager.java
@@ -9,6 +9,7 @@ import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
+import java.util.function.DoubleConsumer;
 import java.util.function.IntConsumer;
 
 public interface Manager {
@@ -176,7 +177,7 @@ public interface Manager {
     @NotNull
     int setWifiConfig(Map<String, String> newConf);
 
-    void addBlockChainProgressListener(@NotNull IntConsumer listener);
+    void addBlockChainProgressListener(@NotNull DoubleConsumer listener);
 
     void uploadNewModule(File inputFile, String fileName) throws ModuleUploadException;
